Use Core's ephemeral server
