+++ packages/test/src/test-workflow-log-interceptor.ts
@@ -11,11 +11,7 @@ const test = anyTest as TestInterface<Context>;
 
 test.before(async (t) => {
   t.context = {
-    testEnv: await TestWorkflowEnvironment.create({
-      testServer: {
-        stdio: 'inherit',
-      },
-    }),
+    testEnv: await TestWorkflowEnvironment.create(),
   };
 });
